Extracted ServerBuildManager
