+++ src/main/java/org/jetbrains/bsp/bazel/server/BepServer.java
@@ -54,12 +54,13 @@ public class BepServer extends PublishBuildEventGrpc.PublishBuildEventImplBase {
   public BepServer(
       BazelBspServer bspServer, BuildClient bspClient, BuildClientLogger buildClientLogger) {
     this.bspServer = bspServer;
-    // TODO BepServer should get BazelData and other parts of the BspServer it needs
-    // TODO in the constructor. To be fixed after the dependency between BepServer and BspServer
-    // TODO is made more sensible
+    // TODO BepServer should get BazelData and BuildManager in the constructor.
+    // TODO To be fixed after the dependency between BepServer and BspServer is made more sensible.
     this.bspClient = bspClient;
     this.buildClientLogger = buildClientLogger;
-    this.diagnosticsDispatcher = new BepDiagnosticsDispatcher(bspServer, bspServer.getBazelData(), bspClient);
+    this.diagnosticsDispatcher =
+        new BepDiagnosticsDispatcher(
+            bspServer.getBazelData(), bspServer.getServerBuildManager(), bspClient);
   }
 
   @Override
